[PC-498] Feat: Social Login Repository, Usecase, Model 생성
