+++ Domain/RepositoryInterfaces/Sources/Login/LoginRepositoryInterfaces.swift
@@ -7,8 +7,7 @@
 
 import SwiftUI
 import Entities
-import DTO
 
 public protocol LoginRepositoryInterfaces {
-  func socialLogin(providerName: SocialLoginType, token: String) async throws -> SocialLoginModel
+  func socialLogin(providerName: SocialLoginType, token: String) async throws -> SocialLoginResultModel
 }
